fix compound key T aspets

+++ SharpRepository.Repository/CompoundKeyRepositoryBase.cs
@@ -25,12 +25,18 @@ namespace SharpRepository.Repository
             get { return _queryManager.CacheUsed; }
         }
 
-        public override IBatch<T> BeginBatch()
+        public ICompoundKeyCachingStrategy<T> CachingStrategy
         {
-            // Return the privately scoped batch via the publicly available interface. 
-            // This ensures that a repository alone can initiate a new batch.
-            return new Batch(this);
+            get { return _cachingStrategy; }
+            set
+            {
+                _cachingStrategy = value ?? new NoCompoundKeyCachingStrategy<T>();
+                _queryManager = new CompoundKeyQueryManager<T>(_cachingStrategy);
+            }
         }
+
+        private readonly RepositoryActionContext<T> _repositoryActionContext;
+        private readonly Dictionary<string, RepositoryActionBaseAttribute> _aspects;
 
         protected CompoundKeyRepositoryBase(ICompoundKeyCachingStrategy<T> cachingStrategy = null)
         {
@@ -41,17 +47,16 @@ namespace SharpRepository.Repository
             RunAspect(aspect => aspect.OnInitialized(_repositoryActionContext));
         }
 
-        private readonly RepositoryActionContext<T> _repositoryActionContext;
-        private readonly Dictionary<string, RepositoryActionBaseAttribute> _aspects;
-
-        public ICompoundKeyCachingStrategy<T> CachingStrategy
+        public override IBatch<T> BeginBatch()
         {
-            get { return _cachingStrategy; }
-            set
-            {
-                _cachingStrategy = value ?? new NoCompoundKeyCachingStrategy<T>();
-                _queryManager = new CompoundKeyQueryManager<T>(_cachingStrategy);
-            }
+            // Return the privately scoped batch via the publicly available interface. 
+            // This ensures that a repository alone can initiate a new batch.
+            return new Batch(this);
+        }
+
+        private void Error(Exception ex)
+        {
+            RunAspect(aspect => aspect.OnError(new RepositoryActionContext<T>(this), ex));
         }
 
         private bool RunAspect(Func<RepositoryActionBaseAttribute, bool> action)
@@ -76,22 +81,86 @@ namespace SharpRepository.Repository
 
         public override IEnumerable<T> GetAll(IQueryOptions<T> queryOptions, IFetchStrategy<T> fetchStrategy)
         {
-            return _queryManager.ExecuteGetAll(
-                () => GetAllQuery(queryOptions, fetchStrategy).ToList(),
-                null,
-                queryOptions
-                );
+            try {
+                var context = new CompoundKeyRepositoryQueryMultipleContext<T>(this, null, queryOptions);
+                if (!RunAspect(attribute => attribute.OnGetAllExecuting(context)))
+                    return context.Results;
+
+                // if the aspect altered the specificaiton then we need to run a FindAll with that specification
+                IEnumerable<T> results;
+                if (context.Specification == null)
+                {
+                    results = _queryManager.ExecuteGetAll(
+                    () => GetAllQuery(queryOptions, fetchStrategy).ToList(),
+                    null,
+                    queryOptions
+                    );
+                } else
+                {
+                    context.Specification.FetchStrategy = fetchStrategy;
+                    results = _queryManager.ExecuteFindAll(
+                        () => FindAllQuery(context.Specification, context.QueryOptions).ToList(),
+                        context.Specification,
+                        null,
+                        context.QueryOptions
+                        );
+                }
+
+                context.Results = results;
+                RunAspect(attribute => attribute.OnGetAllExecuted(context));
+
+                return context.Results;
+            }
+            catch (Exception ex)
+            {
+                Error(ex);
+                throw;
+            }
         }
 
         public override IEnumerable<TResult> GetAll<TResult>(Expression<Func<T, TResult>> selector, IQueryOptions<T> queryOptions, IFetchStrategy<T> fetchStrategy)
         {
-            if (selector == null) throw new ArgumentNullException("selector");
+            try
+            {
+                if (selector == null) throw new ArgumentNullException("selector");
 
-            return _queryManager.ExecuteGetAll(
-                () => GetAllQuery(queryOptions, fetchStrategy).Select(selector).ToList(),
-                selector,
-                queryOptions
-                );
+                var context = new CompoundKeyRepositoryQueryMultipleContext<T, TResult>(this, null, queryOptions, selector);
+                if (!RunAspect(attribute => attribute.OnGetAllExecuting(context)))
+                    return context.Results;
+
+                // if the aspect altered the specificaiton then we need to run a FindAll with that specification
+                IEnumerable<TResult> results;
+
+                if (context.Specification == null)
+                {
+                    results = _queryManager.ExecuteGetAll(
+                        () => GetAllQuery(context.QueryOptions, fetchStrategy).Select(context.Selector).ToList(),
+                        context.Selector,
+                        context.QueryOptions
+                        );
+                }
+                else
+                {
+                    context.Specification.FetchStrategy = fetchStrategy;
+
+                    results = _queryManager.ExecuteFindAll(
+                        () => FindAllQuery(context.Specification, context.QueryOptions).Select(context.Selector).ToList(),
+                        context.Specification,
+                        context.Selector,
+                        context.QueryOptions
+                        );
+                }
+
+                context.Results = results;
+                RunAspect(attribute => attribute.OnGetAllExecuted(context));
+
+                return context.Results;
+            }
+            catch (Exception ex)
+            {
+                Error(ex);
+                throw;
+            }
         }
 
         // These are the actual implementation that the derived class needs to implement
@@ -119,94 +188,207 @@ namespace SharpRepository.Repository
 
         public T Get(params object[] keys)
         {
-            return _queryManager.ExecuteGet(
-                () => GetQuery(keys),
-                null,
-                keys
-                );
+            try
+            {
+                var context = new CompoundKeyRepositoryGetContext<T>(this, keys);
+                if (!RunAspect(attribute => attribute.OnGetExecuting(context)))
+                    return context.Result;
+
+                var result = _queryManager.ExecuteGet(
+                    () => GetQuery(keys),
+                    null,
+                    context.Ids
+                    );
+
+                context.Result = result;
+                RunAspect(attribute => attribute.OnGetExecuted(context));
+
+                return context.Result;
+            }
+            catch (Exception ex)
+            {
+                Error(ex);
+                throw;
+            }
         }
 
         public TResult Get<TResult>(Expression<Func<T, TResult>> selector, params object[] keys)
         {
-            if (selector == null) throw new ArgumentNullException("selector");
+            try
+            {
+                if (selector == null) throw new ArgumentNullException("selector");
 
-            return _queryManager.ExecuteGet(
-                () =>
-                {
-                    var result = GetQuery(keys);
-                    if (result == null)
-                        return default(TResult);
+                var context = new CompoundKeyRepositoryGetContext<T, TResult>(this, keys, selector);
+                if (!RunAspect(attribute => attribute.OnGetExecuting(context)))
+                    return context.Result;
 
-                    var results = new[] { result };
-                    return results.AsQueryable().Select(selector).First();
-                },
-                selector,
-                keys
+                // get the full entity, possibly from cache
+                var result = _queryManager.ExecuteGet(
+                    () =>
+                    {
+                        var singleResult = GetQuery(keys);
+                        if (singleResult == null)
+                            return default(TResult);
+
+                        var results = new[] { singleResult };
+                        return results.AsQueryable().Select(context.Selector).First();
+                    },
+                    context.Selector,
+                    keys
                 );
+
+                context.Result = result;
+                RunAspect(attribute => attribute.OnGetExecuted(context));
+
+                return context.Result;
+            }
+            catch (Exception ex)
+            {
+                Error(ex);
+                throw;
+            }
         }
 
         public override IEnumerable<T> FindAll(ISpecification<T> criteria, IQueryOptions<T> queryOptions = null)
         {
-            if (criteria == null) throw new ArgumentNullException("criteria");
+            if (criteria == null) return GetAll(queryOptions);
 
-            return _queryManager.ExecuteFindAll(
-                () => FindAllQuery(criteria, queryOptions).ToList(),
-                criteria,
-                null,
-                queryOptions
+            try
+            {
+                var context = new CompoundKeyRepositoryQueryMultipleContext<T>(this, criteria, queryOptions);
+                if (!RunAspect(attribute => attribute.OnFindAllExecuting(context)))
+                    return context.Results;
+
+                var results = _queryManager.ExecuteFindAll(
+                    () => FindAllQuery(context.Specification, context.QueryOptions).ToList(),
+                    context.Specification,
+                    null,
+                    context.QueryOptions
                 );
+
+                context.Results = results;
+                RunAspect(attribute => attribute.OnFindAllExecuted(context));
+
+                return context.Results;
+            }
+            catch (Exception ex)
+            {
+                Error(ex);
+                throw;
+            }
+            
         }
 
         public override IEnumerable<TResult> FindAll<TResult>(ISpecification<T> criteria, Expression<Func<T, TResult>> selector, IQueryOptions<T> queryOptions = null)
         {
-            if (criteria == null) throw new ArgumentNullException("criteria");
+            if (criteria == null) return GetAll(selector, queryOptions);
 
-            return _queryManager.ExecuteFindAll(
-                () => FindAllQuery(criteria, queryOptions).Select(selector).ToList(),
-                criteria,
-                selector,
-                queryOptions
-                );
+            try
+            {
+                if (selector == null) throw new ArgumentNullException("selector");
+
+                var context = new CompoundKeyRepositoryQueryMultipleContext<T, TResult>(this, criteria, queryOptions, selector);
+                if (!RunAspect(attribute => attribute.OnFindAllExecuting(context)))
+                    return context.Results;
+
+                var results = _queryManager.ExecuteFindAll(
+                    () => FindAllQuery(context.Specification, context.QueryOptions).Select(context.Selector).ToList(),
+                    context.Specification,
+                    context.Selector,
+                    context.QueryOptions
+                    );
+
+                context.Results = results;
+                RunAspect(attribute => attribute.OnFindAllExecuted(context));
+
+                return context.Results;
+            }
+            catch (Exception ex)
+            {
+                Error(ex);
+                throw;
+            }
         }
         
         public override T Find(ISpecification<T> criteria, IQueryOptions<T> queryOptions = null)
         {
-            if (criteria == null) throw new ArgumentNullException("criteria");
+            try
+            {
+                if (criteria == null) throw new ArgumentNullException("criteria");
 
-            return _queryManager.ExecuteFind(
-                () => FindQuery(criteria, queryOptions),
-                criteria,
-                null,
-                null
-                );
+                var context = new CompoundKeyRepositoryQuerySingleContext<T>(this, criteria, queryOptions);
+                if (!RunAspect(attribute => attribute.OnFindExecuting(context)))
+                    return context.Result;
+
+                var item = _queryManager.ExecuteFind(
+                    () => FindQuery(context.Specification, context.QueryOptions),
+                    context.Specification,
+                    null,
+                    null
+                    );
+
+                context.Result = item;
+                RunAspect(attribute => attribute.OnFindExecuted(context));
+
+                return context.Result;
+            }
+            catch (Exception ex)
+            {
+                Error(ex);
+                throw;
+            }
         }
 
         public override TResult Find<TResult>(ISpecification<T> criteria, Expression<Func<T, TResult>> selector, IQueryOptions<T> queryOptions = null)
         {
-            if (criteria == null) throw new ArgumentNullException("criteria");
-            if (selector == null) throw new ArgumentNullException("selector");
+            try
+            {
+                if (criteria == null) throw new ArgumentNullException("criteria");
+                if (selector == null) throw new ArgumentNullException("selector");
 
-            return _queryManager.ExecuteFind(
-                () =>
-                {
-                    var result = FindQuery(criteria, queryOptions);
-                    if (result == null)
-                        return default(TResult);
+                var context = new CompoundKeyRepositoryQuerySingleContext<T, TResult>(this, criteria, queryOptions, selector);
+                if (!RunAspect(attribute => attribute.OnFindExecuting(context)))
+                    return context.Result;
 
-                    var results = new[] { result };
-                    return results.AsQueryable().Select(selector).First();
-                },
-                criteria,
-                selector,
-                null
-                );
+                var item = _queryManager.ExecuteFind(
+                    () =>
+                    {
+                        var result = FindQuery(context.Specification, context.QueryOptions);
+                        if (result == null)
+                            return default(TResult);
+
+                        var results = new[] { result };
+                        return results.AsQueryable().Select(context.Selector).First();
+                    },
+
+                    context.Specification,
+                    context.Selector,
+                    null
+                    );
+
+                context.Result = item;
+                RunAspect(attribute => attribute.OnFindExecuted(context));
+
+                return context.Result;
+            }
+            catch (Exception ex)
+            {
+                Error(ex);
+                throw;
+            }
         }
 
         public override void Add(T entity)
         {
             if (entity == null) throw new ArgumentNullException("entity");
 
+            if (!RunAspect(attribute => attribute.OnAddExecuting(entity, _repositoryActionContext)))
+                return;
+
             AddItem(entity);
+
+            RunAspect(attribute => attribute.OnAddExecuted(entity, _repositoryActionContext));
+
             if (BatchMode) return;
 
             Save();
@@ -228,7 +410,13 @@ namespace SharpRepository.Repository
         {
             if (entity == null) throw new ArgumentNullException("entity");
 
+            if (!RunAspect(attribute => attribute.OnDeleteExecuting(entity, _repositoryActionContext)))
+                return;
+
             DeleteItem(entity);
+
+            RunAspect(attribute => attribute.OnDeleteExecuted(entity, _repositoryActionContext));
+
             if (BatchMode) return;
 
             Save();
@@ -241,7 +429,13 @@ namespace SharpRepository.Repository
         {
             if (entity == null) throw new ArgumentNullException("entity");
 
+            if (!RunAspect(attribute => attribute.OnUpdateExecuting(entity, _repositoryActionContext)))
+                return;
+
             UpdateItem(entity);
+
+            RunAspect(attribute => attribute.OnUpdateExecuted(entity, _repositoryActionContext));
+
             if (BatchMode) return;
 
             Save();
@@ -310,9 +504,14 @@ namespace SharpRepository.Repository
 
         private void Save()
         {
+            if (!RunAspect(attribute => attribute.OnSaveExecuting(_repositoryActionContext)))
+                return;
+
             SaveChanges();
 
             _queryManager.OnSaveExecuted();
+
+            RunAspect(attribute => attribute.OnSaveExecuted(_repositoryActionContext));
         }
     }
 
@@ -923,5 +1122,6 @@ namespace SharpRepository.Repository
 
             return new Specification<T>(lambda).And(lambda2).And(lambda3);
         }
+
     }
 }
